Added LoginToken Class with generate function
