+++ package.json
@@ -43,6 +43,7 @@
   },
   "devDependencies": {
     "@babel/preset-env": "^7.24.7",
+    "@types/bcryptjs": "^2.4.6",
     "jest": "^29.7.0",
     "nodemon": "^3.1.4",
     "supertest": "^7.0.0"
